3.找出一个薪资最高且公司ID是*的雇员以及该雇员的姓名

+++ src/main/java/com/example/employee/entity/Employee.java
@@ -13,7 +13,7 @@ public class Employee {
     private int age;
     private String gender;
     private int salary;
-    private int companyId;
+    private long companyId;
 
     public Employee() {
     }
@@ -67,11 +67,11 @@ public class Employee {
         this.salary = salary;
     }
 
-    public int getCompanyId() {
+    public long getCompanyId() {
         return companyId;
     }
 
-    public void setCompanyId(int companyId) {
+    public void setCompanyId(long companyId) {
         this.companyId = companyId;
     }
